feature 9th commit

+++ application/views/dash.php
@@ -5,8 +5,10 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="http://localhost/demo/assets/css/dashstyle.css">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/css/bootstrap.min.css" integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
-    
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css" integrity="sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z" crossorigin="anonymous">
+    <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha384-ZvpUoO/+PpLXR1lu4jmpXWu80pZlYUAfxl5NsBMWOEPSjUn/6Z/hRTt8+pR6L4N" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js" integrity="sha384-Q6E9RHvbIyZFJoft+2mJbHaEWldlvI9IOYy5n3zV9zzTtmI3UksdQRVvoxMfooAo" crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js" integrity="sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8sh/joPqPi4lstA6ZIQpPt6wuOM0NoP0OkeagY" crossorigin="anonymous"></script>
     <title>Dashboard</title>
 </head>
 <body>
@@ -27,6 +29,7 @@
       <tr>
        <th>ID</th>
           <th>Tasks</th>
+          <th>Assigned-By</th>
           <th>Start-Date</th>
           <th>End-Date</th>
           <th>Status</th>
@@ -39,12 +42,23 @@
             <tr>
               <td><?php echo $task->sequential;?></td>
               <td><?php echo $task->tname;?></td>
+              <td><?php echo $task->Assignby; ?></td>
               <td><?php echo $task->startD;?></td>
               <td><?php echo $task->endD;?></td>
               <td><?php echo $task->Status;?></td>
               <td>
-                <a href="<?php echo base_url();?>dash/edit/<?php echo $task->id; ?>" class="btn btn-success"  >Edit</a>
-                <a href="<?php echo base_url();?>dash/delete/<?php echo $task->id;?>" class="btn btn-danger" >Delete</a>
+
+              <button type="button" class="edit-btn btn btn-success" 
+                  data-toggle="modal" data-target="#editModal"
+                  data-id="<?php echo $task->id; ?>"
+                  data-taskname="<?php echo $task->tname; ?>"
+                  data-assignby="<?php echo $task->Assignby; ?>"
+                  data-startdate="<?php echo $task->startD; ?>"
+                  data-enddate="<?php echo $task->endD; ?>"
+                  data-status="<?php echo $task->Status; ?>">
+            Edit
+          </button>
+                <a href="<?php echo base_url();?>dash/delete/<?php echo $task->id;?>" class="btn btn-danger" onclick="return confirm('Are you sure you want to delete this task?')" >Delete</a>
               </td>
 
             </tr>
@@ -64,6 +78,10 @@
 
     </div>
     </div>
+    <?php
+// Include edit.php to bring in the modal form
+include 'edittask.php';
+?>
    
   
    
@@ -71,9 +89,9 @@
 <div class="modal fade" id="exampleModal" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel" aria-hidden="true">
   <div class="modal-dialog" role="document">
     <div class="modal-content">
-      <form method="POST" action="<?php echo base_url('dash/addtasks')?>">
+      <form id="addTaskForm"  method="POST" action="<?php echo base_url('dash/addtasks')?>">
       <div class="modal-header">
-        <h5 class="modal-title" id="exampleModalLabel">Modal title</h5>
+        <h5 class="modal-title" id="exampleModalLabel">New Tasks</h5>
         <button type="button" class="close" data-dismiss="modal" aria-label="Close">
           <span aria-hidden="true">&times;</span>
         </button>
@@ -81,15 +99,23 @@
       <div class="modal-body">
       <div class="form-group">
       <label>Task Name:</label>
-      <input type="text" id="task_name" name="task_name" >
+      <input type="text" id="task_name" name="task_name"   >
+      <small id="task_name_error" class="text-danger"></small>
       </div><br>
+      <div class="form-group">
+        <label>Assigned-By</label>
+        <input type="email" id="assign" name="assign">
+        <small id="assign_error" class="text-danger"></small>
+      </div>
       <div class="form-group">
       <label for="start_date">Start Date:</label>
       <input type="date" id="start_date" name="start_date">
+      <small id="start_date_error" class="text-danger"></small>
       </div><br>
       <div class="form-group">
       <label for="end_date">End Date:</label>
       <input type="date" id="end_date" name="end_date">
+      <small id="end_date_error" class="text-danger"></small>
       </div><br>
       <div class="form-group">
       <label for="status">Status:</label>
@@ -98,6 +124,7 @@
         <option value="In-process">In-process</option>
         <option value="Completed">Completed</option>
       </select>
+      <small id="status_error" class="text-danger"></small>
       </div><br>
       </div>
       <div class="modal-footer">
@@ -108,11 +135,23 @@
     </div>
   </div>
 </div>
+
+
+<?php if($this->session->flashdata('error')): ?>
+
+  <div align="center" style="color: #FFF" class="bg-danger">
+  <?php echo $this->session->flashdata('error'); ?>
+  </div>
+<?php endif; ?>
+
+
+
+
+
 <?php if($this->session->flashdata("status")): ?>
   <div align="center" style="color: #FFF" class="bg-success" >
     <?= $this->session->flashdata("status"); ?>
   </div>
-<?php endif; ?>  
-
+<?php endif; ?> 
 </body>
 </html>
